changing MVPA trials (now there 3,4,5TRs waiting times equally spread among correct and wrong positions); fixing MVPA trials (they were in the wrong place, they had unecessary randomization); fixing MVPA bug(missing =last_ttl_timestamp argument in one wait for ttl instruction), using only integers in order not to break MVPA

+++ src/declarativeTask3/ld_utils.py
@@ -67,7 +67,7 @@ def newRandomPresentation(oldPresentation=None, override_remove_cards=None, numb
     else:
         from declarativeTask3.config import removeCards
 
-    newPresentation = np.array(range(matrixSize[0]*matrixSize[1]))
+    newPresentation = np.array(range(matrixSize[0]*matrixSize[1]), dtype=int)
     if len(removeCards):
         removeCards.sort(reverse=True)
         for nCard in removeCards:
